requirements.txt
black==22.1.0
Django==4.0.3
django-allauth==0.49.0
django-crispy-forms==1.14.0
django-filter==21.1
django-shopping-cart==0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
psycopg2-binary==2.9.3
razorpay==1.3.0
waitress==2.1.1
whitenoise==6.0.0
dj-database-url==0.5.0
Pillow==9.0.1
